requirements.txt
openai-agents==0.0.11
deepgram-sdk==3.10.1
fastapi==0.112.0
uvicorn==0.34.0
elevenlabs==1.5.0
python-dotenv==1.0.0 
websockets>=12.0
httpx>=0.21.2
dataclasses-json==0.*
dataclasses==0.*
typing_extensions>=4.0.0,<5.0.0
aenum==3.*
deprecation==2.*
aiohttp>=3.10.2,<4.0.0
aiofiles==23.*
pyaudio>=0.2.14
pydantic>=2.0.0
pydantic-core>=2.18.2
requests>=2.32.2
PyJWT>=2.0.0,<3.0.0
aiohttp-retry==2.8.3
certifi>=2023.7.22
urllib3>=2.2.2
zipp>=3.19.1
python-multipart==0.0.12
webrtcvad==2.0.10
aws-xray-sdk>=2.13
aioredis==2.2.2
asyncpg==0.29.*
sqlalchemy[asyncio]==2.0.*
boto3>=1.34